+++ SW/DataLoggerPC/src/trigger_dialog.h
@@ -28,7 +28,7 @@ public:
 
     ~TriggerDialog();
 
-    static QString prepareFormulaRender(const ConfigFrame* pFrame, const ConfigSignal* pSignal, ConfigTrigger::TriggerCompareOperator oper, unsigned long long rawConstCompareVal);
+    static QString prepareFormulaRender(const ConfigFrame* pFrame, const ConfigSignal* pSignal, ConfigTrigger::TriggerCompareOperator oper, unsigned long rawConstCompareVal);
 private slots:
     void on_signal_comboBox_currentIndexChanged(int index);
     void on_frame_comboBox_currentIndexChanged(int index);
